requirements.txt
scikit-learn
PyYAML
numpy
pandas
streamlit
-e .
python-box #for configbox